feat(apps): web-template starter w/ FSR, Tailwind, vite, monorepo integration & some more
