--- apps/common-web/src/pages/index.tsx
@@ -1,13 +1,0 @@
-import { useState } from 'react'
-import { Button } from 'ui'
-
-function HomePage() {
-  return (
-    <div className="bg-blue-500">
-      <h1 className='underline text-3xl text-white'>zimbas Web</h1>
-      <Button></Button>
-    </div>
-  )
-}
-
-export default HomePage
