+++ apps/web-template/package.json
@@ -1,14 +1,13 @@
 {
-  "name": "common-web",
+  "name": "web-template",
   "private": true,
   "version": "0.0.0",
   "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "preview": "vite preview"
-  },
   "dependencies": {
+    "@tanstack/react-query": "^4.20.4",
+    "@trpc/client": "10.0.0-rc.5",
+    "@trpc/server": "10.0.0-rc.5",
+    "fsr": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.6.1",
@@ -21,8 +20,8 @@
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.20",
     "tailwindcss": "^3.2.4",
+    "tsconfig": "workspace:*",
     "typescript": "^4.6.4",
-    "vite": "^3.2.3",
-    "tsconfig": "workspace:*"
+    "vite": "^3.2.3"
   }
 }
