+++ apps/web-template/src/pages/about.tsx
@@ -1,0 +1,5 @@
+import { Link } from "fsr";
+
+export default function AboutPage() {
+    return <div className="underline">About Page <Link to="/">Get back to Home Page</Link></div>
+}
