+++ apps/web-template/src/vite-env.d.ts
