+++ packages/ui/base/Spinner.tsx
@@ -1,0 +1,4 @@
+import * as React from "react";
+export const Spinner = () => {
+  return <div>Crazy Spinner</div>;
+};
