+++ packages/ui/index.tsx
@@ -1,2 +1,3 @@
 import * as React from "react";
 export * from "./base/Button";
+export * from "./base/Spinner";
